Add function scrapebuildserver()
The build server is dumb and provides no api. If the build fails on the build server we can land in an endless loop waiting for package to be indexed into the repo. Add a scrape function to test whether or not the build has succeeded.

This function still needs to be integrated into the the publish loop.